feat: Add changes requested in PR
