+++ .env.example
@@ -265,3 +265,17 @@ PAYMENTS_SERVICE_BASE_URL=
 PAYMENTS_SERVICE_OAUTH2_CLIENT_ID=
 PAYMENTS_SERVICE_OAUTH2_CLIENT_SECRET=
 PAYMENTS_SERVICE_OAUTH2_SCOPES=payment-profile/read
+
+
+# L5_FORMAT_TO_USE_FOR_DOCS=yaml
+# L5_SWAGGER_GENERATE_ALWAYS=true # Dev setting
+# L5_SWAGGER_GENERATE_YAML_COPY=true
+# L5_SWAGGER_UI_DARK_MODE=true
+
+# Apply a sort to the operation list of each API. It can be 'alpha' (sort by paths alphanumerically), 'method' (sort by HTTP method). Default is the order returned by the server unchanged.
+# L5_SWAGGER_OPERATIONS_SORT=
+# Controls the default expansion setting for the operations and tags. It can be: 'list' (expands only the tags), 'full' (expands the tags and operations), 'none' (expands nothing -default-).
+# L5_SWAGGER_UI_DOC_EXPANSION=none
+# L5_SWAGGER_UI_FILTERS=false # Show/Hide filters. default true
+# L5_SWAGGER_UI_PERSIST_AUTHORIZATION=false # If set to true, it persists authorization data, and it would not be lost on browser close/refresh
+L5_SWAGGER_OPEN_API_SPEC_VERSION=3.1.0
